Revert "feat: adding ack to block produced websocket 🔌"
